# apps/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SummerGear Go API",
    "version": "0.6.0",
    "description": "Go service-session authentication, listings, and private listing image lifecycle. Social provider credentials and Storage credentials may remain disabled in isolated test environments."
  },
  "paths": {
...
    "/api/v1/payments/toss/webhook": {
      "post": {
        "operationId": "paymentWebhook",
        "description": "Untrusted PAYMENT_STATUS_CHANGED hint, deduplicated by provider and transmission ID. Only authoritative PG lookup can update an existing payment attempt. Customer/card payloads are not persisted.",
        "responses": {
          "200": {
            "description": "Event durably received; claimed status is not trusted."
          },
          "400": {
            "description": "Invalid event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Storage or configured payment provider unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "tosspayments-webhook-transmission-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "eventType",
                  "data"
                ],
                "properties": {
                  "eventType": {
                    "const": "PAYMENT_STATUS_CHANGED"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "orderId",
                      "paymentKey",
                      "status"
                    ],
                    "properties": {
                      "orderId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "paymentKey": {
                        "type": "string"
                      },
                      "status": {
                        "enum": [
                          "DONE",
                          "CANCELED",
                          "PARTIAL_CANCELED",
                          "ABORTED",
                          "EXPIRED"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "pattern": "^[a-f0-9]{32}$"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-sg_session"
      }
    }
  }
}